+++ pkg/controller/issueapi/handle_issue_batch.go
